server/pyproject.toml
[project]
name = "robot-buddy-server"
version = "0.1.0"
description = "AI planner server for robot-buddy (runs on 3090 Ti)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.0",
    "websockets>=12.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
llm = [
    "vllm>=0.6",
]
stt = [
    "faster-whisper>=1.0",
]
tts = [
    "orpheus-speech>=0.1",
]
all = [
    "vllm>=0.6",
    "faster-whisper>=1.0",
    "orpheus-speech>=0.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
